+++ src/Manopt.jl
@@ -439,7 +439,7 @@ export AbstractMeshSearchFunction, DefaultMeshAdaptiveDirectSearch
 # Direction Update Rules
 export DirectionUpdateRule
 export Gradient, StochasticGradient
-export AverageGradient, MomentumGradient, Nesterov, PreconditionedGradient
+export AverageGradient, MomentumGradient, Nesterov, PreconditionedDirection
 export SteepestDescentCoefficient,
     HestenesStiefelCoefficient,
     FletcherReevesCoefficient,