+++ Changelog.md
@@ -10,8 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* add a `PreconditionedGradient` variant to the `direction` gradient processor
-  keyword argument and its corresponding `PreconditionedGradientRule`
+* add a `PreconditionedDirection` variant to the `direction` gradient processor
+  keyword argument and its corresponding `PreconditionedDirectionRule`
 
 ### Fixed